backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.29
alembic>=1.13.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
psycopg2-binary>=2.9.0
boto3==1.34.87
httpx==0.27.0
faster-whisper==1.0.3
ollama==0.3.3
transformers==4.40.2
# CPU-only torch: pip install torch --index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
